+++ game/src/devtools/mod.rs
@@ -3,7 +3,7 @@
 
 use abstutil::Timer;
 use geom::{LonLat, Percent};
-use map_gui::common::CityPicker;
+use map_gui::tools::CityPicker;
 use widgetry::{
     lctrl, Btn, Choice, DrawBaselayer, EventCtx, GfxCtx, HorizontalAlignment, Key, Line, Outcome,
     Panel, State, TextExt, VerticalAlignment, Widget,
