+++ game/src/game.rs
@@ -1,3 +1,5 @@
-pub use map_gui::game::{ChooseSomething, PopupMsg, PromptInput};
+// TODO Remove this module entirely
+
+pub use map_gui::tools::{ChooseSomething, PopupMsg, PromptInput};
 
 pub type Transition = widgetry::Transition<crate::app::App>;
