+++ game/src/sandbox/mod.rs
@@ -6,7 +6,7 @@ pub use time_warp::TimeWarpScreen;
 use abstutil::prettyprint_usize;
 use geom::{Circle, Distance, Pt2D, Time};
 use map_gui::load::{FileLoader, MapLoader};
-use map_gui::misc_tools::TurnExplorer;
+use map_gui::tools::TurnExplorer;
 use map_gui::AppLike;
 use sim::{Analytics, Scenario};
 use widgetry::{
