--- map_gui/src/common/mod.rs
@@ -1,13 +1,0 @@
-pub use self::city_picker::CityPicker;
-pub use self::colors::{ColorDiscrete, ColorLegend, ColorNetwork, ColorScale, DivergingScale};
-pub use self::heatmap::{make_heatmap, Grid, HeatmapOptions};
-pub use self::minimap::SimpleMinimap;
-pub use self::navigate::Navigator;
-
-mod city_picker;
-mod colors;
-mod heatmap;
-mod minimap;
-mod navigate;
-#[cfg(not(target_arch = "wasm32"))]
-mod updater;
