+++ map_gui/src/render/parking_lot.rs
@@ -7,9 +7,8 @@ use map_model::{
 use widgetry::{Drawable, EventCtx, GeomBatch, GfxCtx};
 
 use crate::colors::ColorScheme;
-use crate::helpers::ID;
 use crate::render::{DrawOptions, Renderable, OUTLINE_THICKNESS};
-use crate::AppLike;
+use crate::{AppLike, ID};
 
 pub struct DrawParkingLot {
     pub id: ParkingLotID,
