+++ map_gui/src/tools/colors.rs
